Move stage along with the other execution plan implementations
