+++ src/execution_plans/mod.rs
@@ -1,5 +1,7 @@
 mod arrow_flight_read;
 mod partition_isolator;
+mod stage;
 
 pub use arrow_flight_read::ArrowFlightReadExec;
 pub use partition_isolator::{PartitionGroup, PartitionIsolatorExec};
+pub use stage::{display_stage_graphviz, ExecutionTask, StageExec};
