+++ src/protobuf/mod.rs
@@ -1,7 +1,7 @@
 mod distributed_codec;
-mod execution_stage_proto;
+mod stage_proto;
 mod user_codec;
 
 pub(crate) use distributed_codec::DistributedCodec;
-pub(crate) use execution_stage_proto::{proto_from_stage, stage_from_proto, ExecutionStageProto};
+pub(crate) use stage_proto::{proto_from_stage, stage_from_proto, StageExecProto};
 pub(crate) use user_codec::{get_distributed_user_codec, set_distributed_user_codec};
