+++ src/protobuf/stage_proto.rs
@@ -1,7 +1,4 @@
-use std::sync::Arc;
-
-use crate::stage::ExecutionTask;
-use crate::ExecutionStage;
+use crate::execution_plans::{ExecutionTask, StageExec};
 use datafusion::{
     common::internal_datafusion_err,
     error::{DataFusionError, Result},
@@ -12,56 +9,76 @@ use datafusion_proto::{
     physical_plan::{AsExecutionPlan, PhysicalExtensionCodec},
     protobuf::PhysicalPlanNode,
 };
+use std::sync::Arc;
+use url::Url;
 
 #[derive(Clone, PartialEq, ::prost::Message)]
-pub struct ExecutionStageProto {
+pub struct StageExecProto {
     /// Our query id
     #[prost(bytes, tag = "1")]
-    pub query_id: Vec<u8>,
+    query_id: Vec<u8>,
     /// Our stage number
     #[prost(uint64, tag = "2")]
-    pub num: u64,
+    num: u64,
     /// Our stage name
     #[prost(string, tag = "3")]
-    pub name: String,
+    name: String,
     /// The physical execution plan that this stage will execute.
     #[prost(message, optional, boxed, tag = "4")]
-    pub plan: Option<Box<PhysicalPlanNode>>,
+    plan: Option<Box<PhysicalPlanNode>>,
     /// The input stages to this stage
     #[prost(repeated, message, tag = "5")]
-    pub inputs: Vec<ExecutionStageProto>,
+    inputs: Vec<StageExecProto>,
     /// Our tasks which tell us how finely grained to execute the partitions in
     /// the plan
     #[prost(message, repeated, tag = "6")]
-    pub tasks: Vec<ExecutionTask>,
+    tasks: Vec<ExecutionTaskProto>,
+}
+
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct ExecutionTaskProto {
+    /// The url of the worker that will execute this task.  A None value is interpreted as
+    /// unassigned.
+    #[prost(string, optional, tag = "1")]
+    url_str: Option<String>,
+    /// The partitions that we can execute from this plan
+    #[prost(uint64, repeated, tag = "2")]
+    partition_group: Vec<u64>,
 }
 
 pub fn proto_from_stage(
-    stage: &ExecutionStage,
+    stage: &StageExec,
     codec: &dyn PhysicalExtensionCodec,
-) -> Result<ExecutionStageProto, DataFusionError> {
+) -> Result<StageExecProto, DataFusionError> {
     let proto_plan = PhysicalPlanNode::try_from_physical_plan(stage.plan.clone(), codec)?;
     let inputs = stage
         .child_stages_iter()
         .map(|s| proto_from_stage(s, codec))
         .collect::<Result<Vec<_>>>()?;
 
-    Ok(ExecutionStageProto {
+    Ok(StageExecProto {
         query_id: stage.query_id.as_bytes().to_vec(),
         num: stage.num as u64,
         name: stage.name(),
         plan: Some(Box::new(proto_plan)),
         inputs,
-        tasks: stage.tasks.clone(),
+        tasks: stage
+            .tasks
+            .iter()
+            .map(|task| ExecutionTaskProto {
+                url_str: task.url.as_ref().map(|v| v.to_string()),
+                partition_group: task.partition_group.iter().map(|v| *v as u64).collect(),
+            })
+            .collect(),
     })
 }
 
 pub fn stage_from_proto(
-    msg: ExecutionStageProto,
+    msg: StageExecProto,
     registry: &dyn FunctionRegistry,
     runtime: &RuntimeEnv,
     codec: &dyn PhysicalExtensionCodec,
-) -> Result<ExecutionStage> {
+) -> Result<StageExec> {
     let plan_node = msg.plan.ok_or(internal_datafusion_err!(
         "ExecutionStageMsg is missing the plan"
     ))?;
@@ -77,7 +94,7 @@ pub fn stage_from_proto(
         })
         .collect::<Result<Vec<_>>>()?;
 
-    Ok(ExecutionStage {
+    Ok(StageExec {
         query_id: msg
             .query_id
             .try_into()
@@ -86,7 +103,21 @@ pub fn stage_from_proto(
         name: msg.name,
         plan,
         inputs,
-        tasks: msg.tasks,
+        tasks: msg
+            .tasks
+            .into_iter()
+            .map(|task| {
+                Ok(ExecutionTask {
+                    url: task
+                        .url_str
+                        .map(|u| {
+                            Url::parse(&u).map_err(|_| internal_datafusion_err!("Invalid URL: {u}"))
+                        })
+                        .transpose()?,
+                    partition_group: task.partition_group.iter().map(|v| *v as usize).collect(),
+                })
+            })
+            .collect::<Result<Vec<_>>>()?,
         depth: 0,
     })
 }
@@ -96,9 +127,9 @@ pub fn stage_from_proto(
 mod tests {
     use std::sync::Arc;
 
-    use crate::protobuf::execution_stage_proto::ExecutionStageProto;
+    use crate::protobuf::stage_proto::StageExecProto;
     use crate::protobuf::{proto_from_stage, stage_from_proto};
-    use crate::ExecutionStage;
+    use crate::StageExec;
     use datafusion::{
         arrow::{
             array::{RecordBatch, StringArray, UInt8Array},
@@ -147,7 +178,7 @@ mod tests {
             .await?;
 
         // Wrap it in an ExecutionStage
-        let stage = ExecutionStage {
+        let stage = StageExec {
             query_id: Uuid::new_v4(),
             num: 1,
             name: "TestStage".to_string(),
@@ -167,7 +198,7 @@ mod tests {
             .map_err(|e| internal_datafusion_err!("couldn't encode {e:#?}"))?;
 
         // Deserialize from bytes
-        let decoded_msg = ExecutionStageProto::decode(&buf[..])
+        let decoded_msg = StageExecProto::decode(&buf[..])
             .map_err(|e| internal_datafusion_err!("couldn't decode {e:#?}"))?;
 
         // Convert back to ExecutionStage
