--- src/stage/mod.rs
@@ -1,7 +1,0 @@
-mod display;
-mod execution_stage;
-mod task;
-
-pub use display::display_stage_graphviz;
-pub use execution_stage::ExecutionStage;
-pub use task::ExecutionTask;
